Update const.py
I believe this needs to match what's in the manifest. Or at least, I needed to change this in order for the integration config to load.

Note I've gone with simply "idrac_power" as that is what is used in your documentation. If you prefer idrac_power_monitor then we can change the manifest instead.